+++ weight/auxillary_functions.py
@@ -1,19 +1,76 @@
 from datetime import datetime
+import csv
 from sqlalchemy import text, or_, and_
 import sys
-   
+import json
+from flask import jsonify
+
+from classes_db import Container, Transaction, db
+
+def handle_json_in_file(filepath, added):
+    with open(filepath) as jsonfile:
+                    data = json.load(jsonfile)
+                    if not isinstance(data, list):
+                        return jsonify({'error': 'JSON format must be a list'}), 400
+                    for item in data:
+                        cid = item.get('id')
+                        weight = item.get('weight')
+                        unit = item.get('unit', '').lower()
+                        if not cid or weight is None or not unit:
+                            continue
+                        try:
+                            weight = float(weight)
+                        except ValueError:
+                            continue  # Skip invalid weights
+                        existing = Container.query.get(cid)
+                        if existing:
+                            existing.weight = weight
+                            existing.unit = unit
+                        else:
+                            new = Container(container_id=cid, weight=weight, unit=unit)
+                            db.session.add(new)
+                        added += 1
+    return added
+
+def handle_csv_in_file(filepath, added):
+    with open(filepath, newline='') as csvfile:
+                    reader = csv.reader(csvfile)
+                    headers = next(reader, None)  
+                    if not headers or len(headers) < 2:
+                        return jsonify({'error': 'Invalid CSV headers'}), 400
+                    unit = headers[1].lower()  
+                    for row in reader:
+                        if len(row) < 2:
+                            continue  # Skip malformed rows
+                        cid = row[0].strip()
+                        try:
+                            weight = float(row[1])
+                        except ValueError:
+                            continue  # Skip rows with invalid weight
+                        if not cid:
+                            continue  # Skip empty container ID
+                        existing = Container.query.get(cid)
+                        if existing:
+                            existing.weight = weight
+                            existing.unit = unit
+                        else:
+                            new = Container(container_id=cid, weight=weight, unit=unit)
+                            db.session.add(new)
+                        added += 1
+    return added
+
 def get_item_data(date_from, date_to, id):
     default_from = datetime.now().replace(day=1, hour=00, minute=00, second=00, microsecond=00)
     default_to = datetime.now()
 
-    final_from = auxillary_functions.parse_date(date_string=date_from, default_date=default_from)
-    final_to = auxillary_functions.parse_date(date_string=date_to, default_date=default_to)
+    final_from = parse_date(date_string=date_from, default_date=default_from)
+    final_to = parse_date(date_string=date_to, default_date=default_to)
 
     tara, transctions = find_transactions_by_id_and_time(id=id, start_time=final_from, end_time=final_to)
     if transctions is None:
         transctions = find_transactions_by_container(container_id=id, start_time=final_from, end_time=final_to)
         container: Container = db.session.query(Container).filter(Container.container_id == id).first()
-        unit, tara = auxillary_functions.lb_to_kg(unit=container.unit, weight=container.weight)
+        unit, tara = lb_to_kg(unit=container.unit, weight=container.weight)
     
     if transctions is None:
         return f"ID {id} not found in requested time range.", 404
@@ -21,16 +78,10 @@ def get_item_data(date_from, date_to, id):
     session_ids = []
     for transaction in transctions:
         session_ids.append(transaction.session_id)
-    auxillary_functions.print_debug(f"len of sessions id list at first: {len(session_ids)}")
-    auxillary_functions.print_debug(f"session id list is: {session_ids}")
     list(set(session_ids))
-    auxillary_functions.print_debug(f"after set, list len is {len(session_ids)}")
-    auxillary_functions.print_debug(f"after set, list is {session_ids}")
     for _item in session_ids:
         if _item is None:
             session_ids.remove(_item)
-    auxillary_functions.print_debug(f"after None removal, list len is {len(session_ids)}")
-    auxillary_functions.print_debug(f"after None remova, list is {session_ids}")
     
     ret = {"id": id, "tara": tara if tara else 'na', 'sessions': session_ids, 'unit': 'kg'}
     return ret
@@ -198,4 +249,79 @@ def get_transactions_by_time_range(db_session,Transaction_model,from_time, to_ti
         print(f"Error executing query: {e}")
         return []
     
+class truck_direction():
+
+    def truck_in(data: json):
+        session_id = None
+        already_exists = False
+        try:
+            prev_record = json.loads(data['prev_record'])
+            if prev_record['direction'] == 'in':
+                if not data['force']:
+                    return 'Bad Request', 400
+                elif prev_record['truck'] != data['truck']:
+                        return 'Bad Request', 400 # better text, unsure can happen
+                else:
+                    session_id = prev_record['session_id']
+                    already_exists = True
+        except KeyError:
+            session_id = create_session_id(data['datetime'])
+        data['unit'], data['weight'] = lb_to_kg(data['unit'], data['weight'])
+        data['bruto'] = data['weight']
+        new_transaction = create_transaction_from_data_and_session_id(data=data, session_id=session_id)
+        insert_transaction(new_transaction=new_transaction, exists=already_exists)
+        ret_json = {'id': new_transaction.id, 'truck': new_transaction.truck, 'bruto': new_transaction.bruto}
+        return ret_json, 200
+            
+
+    def truck_out(data: json):
+        try:
+            entrance = json.loads(data['prev_record'])
+        except KeyError:
+            return "No in for this out", 400
+        if entrance['direction'] == 'out':
+            if not data['force']:
+                    return 'Two outs is a row without force', 400
+            elif entrance['truck'] != data['truck']:
+                return 'Bad Request', 400 # better text
+        bruto = entrance['bruto']
+        truck_tara = data['weight']
+        containers_tara = 0
+        container_ids = entrance['containers']
+        containers = db.session.query(Container).filter(Container.container_id.in_(container_ids)).all()
+        for container in containers:
+            if isinstance(container.weight, int) and container.weight > 0:
+                container.unit, container.weight = lb_to_kg(container.unit, container.weight)
+                containers_tara += container.weight
+            else:
+                containers_tara = 'NA'
+                break
+        neto = bruto - truck_tara - containers_tara
+        id = create_session_id(data['datetime']) #this is also the id - not session id - for out
+        new_transaction = in_json_and_extras_to_transaciotn(in_json=entrance, truck_tara=truck_tara, neto=neto, exact_time=data['datetime'], id=id)
+        db.session.add(new_transaction)
+        db.session.commit()
+        ret = {'id': new_transaction.id, 'truck': new_transaction.truck, 'bruto': new_transaction.bruto, 'truckTara': new_transaction.truckTara, 'neto': new_transaction.neto}
+        return ret, 200
     
+    def truck_none(data: json):
+                # none after in should generate an error. Why? What does it mean?
+        new_tansaction = Transaction()
+        new_tansaction.bruto = data['weight']
+        container_id = data['containers'][0] # אthis implementation of none only accepts single container
+        container = containers = db.session.query(Container).filter_by(container_id=container_id).first()
+        unit, container_tara = lb_to_kg(container.unit ,container.weight)
+        new_tansaction.truckTara = container_tara #should this be the case?
+        new_tansaction.containers = [container]
+        new_tansaction.neto = new_tansaction.bruto - container_tara
+        new_tansaction.id = new_tansaction.session_id = create_session_id(data['datetime'])
+        new_tansaction.direction = 'none'
+        new_tansaction.produce = data['produce']
+        new_tansaction.truck = 'na'
+        new_tansaction.datetime = data['datetime']
+        db.session.add(new_tansaction)
+        db.session.commit()
+        ret = {'id': new_tansaction.id, 'truck': new_tansaction.truck, 'bruto': new_tansaction.bruto, 'truckTara': new_tansaction.truckTara, 'neto': new_tansaction.neto}
+        return ret, 200
+    
+truck_direction_handler = {'in': truck_direction.truck_in, 'out': truck_direction.truck_out, 'none': truck_direction.truck_none}
